+++ lib/core/system/configuration.dart
@@ -3,6 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:lunasea/core.dart';
 
 class LunaConfiguration {
+  static const String _EXTERNAL_MODULES_KEY = 'external_modules';
+  static const String _INDEXERS_KEY = 'indexers';
+  static const String _LUNASEA_KEY = 'lunasea';
+  static const String _PROFILES_KEY = 'profiles';
+
   /// Returns a list of all profiles converted to a map.
   List<Map<String, dynamic>> _getProfiles() {
     List<Map<String, dynamic>> _data = [];
@@ -32,10 +37,24 @@ class LunaConfiguration {
     for (Map indexer in data) box.add(IndexerHiveObject.fromMap(indexer));
   }
 
+  /// Returns a list of all external modules converted to a map.
+  List<Map<String, dynamic>> _getExternalModules() {
+    List<Map<String, dynamic>> _data = [];
+    for (var key in Database.externalModulesBox.keys)
+      _data.add(Database.externalModulesBox.get(key).toMap());
+    return _data;
+  }
+
+  /// Given a list of map objects, creates or updates external modules for each object.
+  void _setExternalModules(List data) {
+    Box<dynamic> box = Database.externalModulesBox;
+    for (Map module in data) box.add(ExternalModuleHiveObject.fromMap(module));
+  }
+
   /// Import the entire configuration from a JSON-encoded string (typically read through a `.lunasea` backup file).
   ///
   /// - Clears all boxes
-  /// - Calls `_setProfiles()` and `_setIndexers()`
+  /// - Calls [_setProfiles], [_setIndexers], and [_setExternalModules].
   /// - Calls `import()` on all module databases, which implement [LunaDatabase].
   /// - Resets the application state
   ///
@@ -44,16 +63,22 @@ class LunaConfiguration {
     Map config = json.decode(data);
     Database().clearAllBoxes();
     try {
-      if (config['profiles'] != null) _setProfiles(config['profiles']);
-      if (config['indexers'] != null) _setIndexers(config['indexers']);
-      if (config['lunasea'] != null) LunaDatabase().import(config['lunasea']);
+      if (config[_PROFILES_KEY] != null) _setProfiles(config[_PROFILES_KEY]);
+      if (config[_INDEXERS_KEY] != null) _setIndexers(config[_INDEXERS_KEY]);
+      if (config[_EXTERNAL_MODULES_KEY] != null)
+        _setExternalModules(config[_EXTERNAL_MODULES_KEY]);
+      if (config[_LUNASEA_KEY] != null)
+        LunaDatabase().import(config[_LUNASEA_KEY]);
       LunaModule.values.forEach((module) {
         if (config[module.key] != null)
           module.database?.import(config[module.key]);
       });
     } catch (error, stack) {
       LunaLogger().error(
-          'Failed to import configuration, resetting to default', error, stack);
+        'Failed to import configuration, resetting to default',
+        error,
+        stack,
+      );
       Database().setDefaults();
     }
     // Reset the entire app's state
@@ -66,9 +91,10 @@ class LunaConfiguration {
   /// - Calls `export()` on all module databases, which implement [LunaModuleDatabase].
   String export() {
     Map<String, dynamic> config = {
-      "profiles": _getProfiles(),
-      "indexers": _getIndexers(),
-      "lunasea": LunaDatabase().export(),
+      _EXTERNAL_MODULES_KEY: _getExternalModules(),
+      _INDEXERS_KEY: _getIndexers(),
+      _PROFILES_KEY: _getProfiles(),
+      _LUNASEA_KEY: LunaDatabase().export(),
     };
     LunaModule.values.forEach((module) {
       if (module.database != null)