+++ lib/modules/external_modules/routes/external_modules.dart
@@ -1,0 +1,2 @@
+export 'external_modules/route.dart';
+export 'external_modules/widgets.dart';