+++ lib/modules/external_modules/routes/external_modules/widgets.dart
@@ -1,0 +1,1 @@
+export 'widgets/module_tile.dart';