+++ lib/modules/external_modules/routes/external_modules/widgets/module_tile.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:lunasea/core.dart';
+
+class ExternalModulesModuleTile extends StatelessWidget {
+  final ExternalModuleHiveObject module;
+
+  ExternalModulesModuleTile({
+    Key key,
+    @required this.module,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return LunaListTile(
+      context: context,
+      title: LunaText.title(text: module.displayName),
+      subtitle: LunaText.subtitle(text: module.host),
+      trailing: LunaIconButton(icon: Icons.arrow_forward_ios_rounded),
+      onTap: () async => module.host.lunaOpenGenericLink(),
+    );
+  }
+}