+++ lib/modules/settings/core/router.dart
@@ -19,6 +19,10 @@ class SettingsRouter extends LunaModuleRouter {
     // Configuration/Dashboard
     SettingsConfigurationDashboardRouter().defineRoute(router);
     SettingsConfigurationDashboardCalendarSettingsRouter().defineRoute(router);
+    // Configuration/External Modules
+    SettingsConfigurationExternalModulesRouter().defineRoute(router);
+    SettingsConfigurationExternalModulesAddRouter().defineRoute(router);
+    SettingsConfigurationExternalModulesEditRouter().defineRoute(router);
     // Configuration/Lidarr
     SettingsConfigurationLidarrRouter().defineRoute(router);
     SettingsConfigurationLidarrConnectionDetailsRouter().defineRoute(router);