+++ lib/modules/settings/routes/configuration_externalmodules.dart
@@ -1,0 +1,2 @@
+export 'configuration_externalmodules/pages.dart';
+export 'configuration_externalmodules/route.dart';