+++ lib/modules/settings/routes/configuration_externalmodules/pages.dart
@@ -1,0 +1,2 @@
+export 'pages/add_module.dart';
+export 'pages/edit_module.dart';